requirements.txt
streamlit~=1.37.0
numpy~=2.0.1
pillow~=10.4.0
scikit-image~=0.24.0
opencv-python-headless